+++ extension/webview/src/components/index.ts
@@ -10,6 +10,7 @@ export { EmptyState } from './EmptyState';
 export { Tooltip } from './Tooltip';
 export { Welcome } from './Welcome';
 export { Logo } from './Logo';
+export { Markdown } from './Markdown';
 export { 
     AttachmentInput, 
     AttachmentBadge,